pyproject.toml
[project]
name = "offrl-preset"
version = "0.1.0"
description = "Local workspace preset for offline RL research (not oriented toward PyPI install)"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "torch>=1.13",
    "gym>=0.21",
    "pyyaml>=6.0",
    "d4rl",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["offrl*", "algorithms*"]

[tool.setuptools.package-data]
offrl = ["py.typed"]
